RabbitCarrotGame-Java
Multithreaded Java program for the following game. We have some boxes arranged sequentially in the game. We have several rabbits and they jump from one box to another. There is someone who puts carrots in the boxes for the rabbits, and the rabbits will eat these carrots and move on to the last box.